requirements.txt
numpy~=1.26.4
tensorflow~=2.16.2
streamlit
pillow
tensorflow